Diet-Plan-Generator-main/Diet-Plan-Generator-main/DietPlanner/requirements.txt
streamlit==1.32.0
langchain==0.1.5
langchain-groq==0.1.2
python-dotenv==1.0.0